feat(history): define comment history model
In order to track comment history, we would need to have a copy of
the comment on each edit. Also for future queries, we've added a column
"isEdited" which is a boolean to know if the comment has been edited before and
to determine whether or not to fetch it's history.
- add isEdited column to comment table
- generate commentHistory table and migrations
- add relationship between both tables