#159987737 Comments should be able to track edit history (#47)
* feat(comments): add ability to get all comments/replies of an article

Fetch all comments and replies for an article and include user informantion.
- update comment models: add validations.

* feat(history): define comment history model

In order to track comment history, we would need to have a copy of
the comment on each edit. Also for future queries, we've added a column
"isEdited" which is a boolean to know if the comment has been edited before and
to determine whether or not to fetch it's history.
- add isEdited column to comment table
- generate commentHistory table and migrations
- add relationship between both tables

* feat(comment): track comment edit history functionality

Allow users update, their comments, and also track the edit history
of each update.
- add updateComment method
- prevent users from updating other users' comments
- add relevant tests
- use hooks to update commentHistory model
- return updated comment with each update

* feat(history): add method to retrieve comment edit history

Users can comment and edit comments. This method fetches the comment edit history.
- Add relevant tests
- Add getCommentEdits method.
[Delivers #159987737]